fundingrate/requirements.txt
streamlit
pandas
numpy
requests
plotly
tenacity
python-dateutil
